requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
Cython==0.29.20
Flask==1.1.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.1.2
pytesseract==0.3.4
requests==2.23.0
tesseract==0.1.3
tesseract-ocr==0.0.1
urllib3==1.25.9
Werkzeug==1.0.1
